Apply minor suggestions from the feedback

+++ cpp/src/io/parquet/experimental/hybrid_scan_preprocess.cu
@@ -180,31 +180,31 @@ hybrid_scan_reader_impl::prepare_dictionaries(
         options.get_timestamp_type().id(),
         schema.logical_type);
 
-      // Create a column chunk
+      // Create a column chunk descriptor - zero/null values for all fields that are not needed
       chunks[chunk_idx] = ColumnChunkDesc(static_cast<int64_t>(dict_page_data.size()),
                                           static_cast<uint8_t*>(dict_page_data.data()),
                                           col_meta.num_values,
                                           schema.type,
                                           schema.type_length,
-                                          0,  // Not needed
-                                          0,  // Not needed
-                                          0,  // Not needed
-                                          0,  // Not needed
-                                          0,  // Not needed
-                                          0,  // Not needed
-                                          0,  // Not needed
+                                          0,  // start_row
+                                          0,  // num_rows
+                                          0,  // max_definition_level
+                                          0,  // max_repetition_level
+                                          0,  // max_nesting_depth
+                                          0,  // def_level_bits
+                                          0,  // rep_level_bits
                                           col_meta.codec,
                                           schema.logical_type,
                                           clock_rate,
-                                          0,  // Not needed
+                                          0,  // src_col_index
                                           col_schema_idx,
-                                          nullptr,  // Not needed
-                                          0.0f,     // Not needed
-                                          false,    // Not needed
+                                          nullptr,  // chunk_info
+                                          0.0f,     // list_bytes_per_row_est
+                                          false,    // strings_to_categorical
                                           rg.source_index);
       // Set the number of dictionary and data pages
-      chunks[chunk_idx].num_dict_pages = (dict_page_data.size() > 0);
-      chunks[chunk_idx].num_data_pages = 0;
+      chunks[chunk_idx].num_dict_pages = static_cast<int32_t>(dict_page_data.size() > 0);
+      chunks[chunk_idx].num_data_pages = 0;  // Always zero at this stage
       chunk_idx++;
     }
   }
